Added Product Categories Page, Table View and Input Field for Product
You can create categories edit or delete them and assign them to products, only issue is that we can't currently expand products along with category using useAdminCategory like with collections so I decided to leave it empty. So in the add-product-table you won't see an indication that shows which products are related to the category. However, you can still see them in the view-products-table